RGB MNIST and USPS, ready for PR
